requirements.txt
awscli==1.10.58
boto3==1.4.0
botocore==1.4.48
colorama==0.3.7
docutils==0.12
futures==3.0.5
jmespath==0.9.0
Pantry==0.3
picamera==1.12
pyasn1==0.1.9
python-dateutil==2.5.3
PyYAML==3.11
rsa==3.4.2
s3transfer==0.1.2
six==1.10.0
